OSD-4470 custom-operator-metrics
This PR makes use of the
https://github.com/openshift/operator-custom-metrics repo
to implement specific custom metrics for MUO. I had to `replace` the
custom metrics repo with my fork to support the required `svc` labels.
We can repoint this back after
https://github.com/openshift/operator-custom-metrics/pull/37 is merged.

Metric names were also refactored according to:
* https://prometheus.io/docs/practices/naming/
* https://docs.google.com/document/d/1rCKAYTrYMESjJDyJ0KvNap05NNukmNXVVi6MShISnOw/edit#
* https://github.com/kubernetes/community/blob/master/contributors/devel/sig-instrumentation/instrumentation.md#naming
* https://www.robustperception.io/target-labels-not-metric-name-prefixes

This PR will be the first of a few to refactor our metrics/alerting into
these practices. Ideally we should combined the metrics and collector
packages in future PRs.