OSD-11002 patch clusterversion instead of update
An Update() to clusterversion can result in potentially cause fields to be nullified or emptied in it,
if the client performing the update isn't aware of those fields.
This will impact MUO as it performs an Update to clusterversion when it initiates an upgrade.
Instead of an Update it should perform a Patch.