pkg/upgraders/upgradeable: Pass through the ClusterVersion Upgradeable message
Instead of assuming we know what the issue is, just bubble up the
issue as explained by the cluster-version operator.  This thins out
the update stack, simplifying maintenance, and avoids
mischaracterizing issues like this ClusterVersion condition:

  - lastTransitionTime: "2022-11-18T00:52:52Z"
    message: |-
      Cluster operator monitoring should not be upgraded between minor versions: Highly-available workload in namespace openshift-monitoring, with label map["app.kubernetes.io/name":"prometheus"] and persistent storage enabled has a single point of failure.
      Highly-available workload in namespace openshift-monitoring, with label map["app.kubernetes.io/name":"alertmanager"] and persistent storage enabled has a single point of failure.
      Manual intervention is needed to upgrade to the next minor version. For each highly-available workload that has a single point of failure please mark at least one of their PersistentVolumeClaim for deletion by annotating them with map["openshift.io/cluster-monitoring-drop-pvc":"yes"].
    reason: WorkloadSinglePointOfFailure
    status: "False"
    type: Upgradeable

as an "unacknowledged user actions" issue.

One complication is that responsibility for resolving
Upgradeable=False is shared between customers (e.g. AdminAckRequired)
and service-delivery (e.g. WorkloadSinglePointOfFailure), but we can
mitigate that by proactively monitoring for and resolving reasons that
fall on the service-delivery side of things.